Clean up Route & WingitUser packages - Added new deleteUser_NotFound_Failure test case - Removed stack trace for production from rest exception handling - Cleaned up and standardized towards class activity style of functions and responses.

+++ wingit/src/main/java/com/G2T5203/wingit/user/UserService.java
@@ -2,19 +2,12 @@ package com.G2T5203.wingit.user;
 
 import com.G2T5203.wingit.entities.WingitUser;
 import jakarta.transaction.Transactional;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class UserService {
-    private static final Logger logger = LoggerFactory.getLogger(UserService.class);
-
     private final UserRepository repo;
 
     public UserService(UserRepository repo) {
@@ -30,49 +23,21 @@ public class UserService {
     }
 
     @Transactional
-    public HttpStatus createUser(WingitUser newUser) {
-        try {
-            repo.save(newUser);
-        } catch (DataIntegrityViolationException e) {
-            // TODO: Convert these systems to custom Exceptions based. Refer to class activity BookExceptions.
-            logger.error("Failed to add new User: DataIntegrityViolationException\n" + newUser.toString());
-            logger.debug("Error details: " + e.getLocalizedMessage());
-            return HttpStatus.BAD_REQUEST;
-        }
-
-        return HttpStatus.CREATED;
-    }
+    public WingitUser createUser(WingitUser newUser) { return repo.save(newUser); }
 
     @Transactional
-    public HttpStatus deleteUserById(Integer userId) {
-        try {
+    public void deleteUserById(Integer userId) {
+        if (repo.existsById(userId)) {
             repo.deleteById(userId);
-        } catch (EmptyResultDataAccessException e) {
-            logger.error("Failed to delete user by Id: EmptyResultDataAccessException\n" +
-                    "UserId that was attempted to be deleted: " + userId);
-            logger.debug("Error details: " + e.getLocalizedMessage());
-            return HttpStatus.BAD_REQUEST;
+        } else {
+            throw new UserNotFoundException(userId);
         }
-
-        return HttpStatus.OK;
     }
 
     @Transactional
-    public HttpStatus updateUser(WingitUser updatedUser) {
-        try {
-            boolean userExists = getById(updatedUser.getUserId()) != null;
-            if (userExists) {
-                repo.save(updatedUser);
-            } else {
-                logger.error("User " + updatedUser.getUserId() + " does not exist!");
-                return HttpStatus.BAD_REQUEST;
-            }
-        } catch (DataIntegrityViolationException e) {
-            logger.error("Failed to add new User: DataIntegrityViolationException\n" + updatedUser.toString());
-            logger.debug("Error details: " + e.getLocalizedMessage());
-            return HttpStatus.BAD_REQUEST;
-        }
-
-        return HttpStatus.OK;
+    public WingitUser updateUser(WingitUser updatedUser) {
+        boolean userExists = repo.existsById(updatedUser.getUserId());
+        if (!userExists) throw new UserNotFoundException(updatedUser.getUserId());
+        return repo.save(updatedUser);
     }
 }
